openCV/requirements.txt
opencv-python==4.13.0.92
numpy==2.1.1
streamlit==1.41.1
pillow==11.1.0
